refactor: merge dev and build tool files
